LPS-120269 Move mapper to apps/petra
