+++ modules/apps/petra/petra-url-pattern-mapper/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Petra URL Pattern Mapper
+Bundle-SymbolicName: com.liferay.petra.url.pattern.mapper
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.petra.url.pattern.mapper
