+++ modules/apps/petra/petra-url-pattern-mapper/build.gradle
@@ -1,0 +1,3 @@
+liferay {
+	deployDir = appServerLibGlobalDir
+}
