+++ modules/apps/petra/petra-url-pattern-mapper/src/main/resources/com/liferay/petra/url/pattern/mapper/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
